moving handler to delete package
